adk-sock-shop/pyproject.toml
[project]
name = "adk-demo"
version = "0.1.0"
description = "ADK agent demo with Docker Compose"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.5.0",
    "litellm>=1.73.2",
    "streamlit>=1.31.0",
    "requests>=2.31.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "sseclient-py>=1.8.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.402",
    "ruff>=0.12.1",
]
